requirements.txt
# HomeworkHelper 의존성

# === GUI ===
PyQt6

# === Windows ===
pywin32
winshell
psutil

# === API Server ===
fastapi
uvicorn
SQLAlchemy
pydantic

# === Network ===
requests

# === Data Validation ===
jsonschema

# === Windows Notifications ===
Windows-Toasts

# === HoYoLab Integration ===
genshin>=1.6.0
pycryptodome>=3.20.0

# === Build/Dev ===
pyinstaller
pytest
